airflow/requirements-airflow.txt
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
pyspark==3.5.0
boto3==1.34.0
apache-airflow-providers-amazon==8.13.0
apache-airflow-providers-apache-spark==4.3.0